+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+0.16.0
+==================
+
+* Add named groups
+* `match` returns a `Match` object now
+
+0.15.0
+==================
+
+* Add `\A`, `\z`, `\B` assertions
+
+0.14.0
+==================
+
+* Add lookahead assertion `(?=...)` and `(?!...)` (limited to a single char)
+
+0.13.0
+==================
+
+* Add set complement `[^...]`
+
+0.12.0
+==================
+
+* Add `\s`, `\W`, `\D`, `\S` shorthands
+
+0.11.0
+==================
+
+* Add `.` match any
+
+0.10.0
+==================
+
+* Add `\b` word boundary (empty match)
+
+0.9.0
+==================
+
+* Add start `^` and end `$` (empty matches)
+
 0.8.0
 ==================